+++ CHANGELOG.md
@@ -5,13 +5,15 @@ This project follows [Semantic Versioning](http://semver.org/).
 
 ## [1.15.0] - 2020-10-30
 ### Added
-- Build/Load Cohort(Instead of Patient selection)
+- Build/Load Cohort (instead of Patient selection)
 ## Changed
-- Filter Patients by Observation codes if no value is provided for them
+- Allow patients to be selected by Observation codes which do not have
+  values provided for them
 - Changed the Patient.active parameter input control from a checkbox to
   a radio button group
 ## Fixed
 - Exception that can be thrown after removing a resource type tab
+- Results of not all queries were considered when filtering patient resources
 
 ## [1.14.0] - 2020-10-16
 ### Added